Placed App component in its own file
