+++ package.json
@@ -15,13 +15,9 @@
     "test": "jest"
   },
   "dependencies": {
-    "@babel/core": "^7.8.7",
-    "@babel/preset-react": "^7.8.3",
-    "babel-loader": "^8.0.6",
     "bluebird": "^3.7.2",
     "css-loader": "^3.4.2",
     "express": "^4.17.1",
-    "jest": "^25.1.0",
     "lodash": "^4.17.15",
     "mysql": "^2.18.1",
     "react": "^16.13.0",
@@ -30,7 +26,22 @@
     "webpack": "^4.42.0"
   },
   "devDependencies": {
+    "@babel/core": "^7.8.7",
+    "@babel/preset-env": "^7.9.0",
+    "@babel/preset-react": "^7.8.3",
+    "babel-jest": "^25.1.0",
+    "babel-loader": "^8.0.6",
+    "enzyme": "^3.11.0",
+    "enzyme-adapter-react-16": "^1.15.2",
     "eslint-config-hackreactor": "git://github.com/reactorcore/eslint-config-hackreactor",
+    "identity-obj-proxy": "^3.0.0",
+    "jest": "^25.1.0",
+    "react-test-renderer": "^16.13.1",
     "webpack-cli": "^3.3.11"
+  },
+  "jest": {
+    "moduleNameMapper": {
+      "\\.(css|less)$": "identity-obj-proxy"
+    }
   }
 }
