+++ public/index.html
@@ -7,7 +7,9 @@
 </head>
 <body>
     <div id="root"></div>
-    <script src="bundle.js"></script>
+
+    <script src="dist/bundle.js"></script>
+    <!-- let's check if this is the same as localhost: etc. -->
 </body>
 </html>
