+++ webpack.config.js
@@ -1,9 +1,9 @@
 const path = require('path');
 const SRC_DIR = path.join(__dirname, '/client/src');
-const DIST_DIR = path.join(__dirname, '/client/dist');
+const DIST_DIR = path.join(__dirname, '/public/dist');
 
 module.exports = {
-    //mode: "development",
+    mode: "development",
     //devtool: 'none', 
     entry: `${SRC_DIR}/index.jsx`,
     output: {
@@ -15,9 +15,10 @@ module.exports = {
             {
                 test: /\.jsx$/,
                 loader: 'babel-loader',
-                exclude: ["/node_modules"],
+                exclude: /node_modules/,
                 options: {
-                        presets: ['@babel/preset-react']
+                        presets: ['@babel/preset-react', 
+                                  '@babel/preset-env']
                 }
             },
             {
@@ -26,5 +27,6 @@ module.exports = {
             }
         ]
     }
+
 
 }
